+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 
-- Add `PrometheusMetrics::get_http_requests_total` function.
+- Add `PrometheusMetrics::http_requests_total` and `PrometheusMetrics::http_requests_duration_seconds` methods to get a reference to the fairing's internal Prometheus metrics. 
 - Add `PrometheusMetrics::get_http_requests_duration_seconds` function.
 
 ## [0.10.0-rc.2] - 2022-05-14